requirements.txt
fastapi
uvicorn
pydantic
PyYAML
mysql-connector-python
